refactoring, remove app.js, move functions to chef and desk
